Modify Search and Filter based on new design
