+++ src/components/Button/components/Secondary.tsx
@@ -1,0 +1,15 @@
+import Button from "../base";
+import { IButton } from "../base";
+
+function ButtonSecondary({ children, ...props }: IButton) {
+  return (
+    <Button
+      {...props}
+      className={["Vlt-btn--secondary"]}
+    >
+      {children}
+    </Button>
+  )
+}
+
+export default ButtonSecondary;
