+++ src/components/Button/index.tsx
@@ -1,9 +1,12 @@
 import ButtonPrimary from "./components/Primary";
+import ButtonSecondary from "./components/Secondary";
 import ButtonTertiary from "./components/Tertiary";
 
 function Button() {}
 
 Button.Primary = ButtonPrimary;
+Button.Secondary = ButtonSecondary;
 Button.Tertiary = ButtonTertiary;
+
 
 export default Button;
